requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python-headless==4.10.0.84
Pillow>=9.0.0
matplotlib>=3.5.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Web interface dependencies
streamlit>=1.25.0
plotly>=5.10.0

# Image processing
scikit-image>=0.19.0

# Document generation
python-docx>=0.8.11
reportlab>=3.6.0
PyMuPDF>=1.23.0
lxml>=4.9.0

# Visualization
seaborn>=0.11.2

# Additional dependencies
exifread>=3.0.0
imageio>=2.31.0

# Optional: For PDF conversion
# python-libreoffice  # For PDF conversion (optional, requires LibreOffice installed)
